methods/contVid/requirements.txt
accelerate==0.17.1
addict==2.4.0
basicsr==1.4.2
bitsandbytes==0.35.4
clip==0.1.0
cmake==3.25.2
controlnet-aux==0.0.6
decord==0.6.0
deepspeed==0.8.0
diffusers==0.14.0
easydict==1.10
einops==0.6.0
ffmpy==0.3.0
ftfy==6.1.1
imageio==2.25.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.7.0.68
pandas==1.5.3
pillow==9.4.0
scikit-image==0.19.3
scipy==1.10.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
thinc==8.1.10
timm==0.6.12
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
wandb==0.13.10
xformers==0.0.16
modelcards
